Code-review 2026-07-11 §6.2 (move 6): split sql_gen.rs tests into behaviour-named files
sql_gen.rs was ~490 lines of production code buried under ~3,770 lines of a
single phase-named `mod tests` (16 flat nested modules, several still named
`phaseNN_*`). The review calls out that this phase-named archaeology is what
made E-1's alias-shadowing blind spot invisible.

Extract all 16 nested test modules into behaviour-named sibling files under
src/expand/ (tests_fan_trap.rs, tests_derived_metric.rs, tests_role_playing.rs,
…), dropping the phaseNN_ prefixes. sql_gen.rs drops from 4,259 to 491 lines
(production only). Test bodies move verbatim — `use super::*` now resolves
against crate::expand's re-exports (the entry point + request/error types the
tests use), which are the same names the old `mod tests` imported.

No test lost or renamed: the 125 extracted tests all pass and the full suite is
unchanged at 1155 (sqllogictest 71/0, clippy default + extension clean, build
green). Tests-only change — no production behavior touched.
